// src/lib/duckdb-serialization.gen.ts
export type DuckDBValue = unknown;
export type DuckDBLogicalType = unknown;
// ...
export interface AtClause {
  unit: string;
  expr: AnyParsedExpression | null;
}
// ...
export interface BaseTableRef extends TableRef {
  schema_name: string;
  table_name: string;
  column_name_alias: string[];
  catalog_name: string;
  at_clause: AtClause | null;
  type: "BASE_TABLE";
}
export interface BetweenExpression extends ParsedExpression {
  input: AnyParsedExpression | null;
  lower: AnyParsedExpression | null;
  upper: AnyParsedExpression | null;
  class: "BETWEEN";
}
// ...
export interface CTENode extends QueryNode {
  cte_name: string;
  query: AnyQueryNode | null;
  child: AnyQueryNode | null;
  aliases: string[];
  materialized: string;
  type: "CTE_NODE";
}
export interface CaseCheck {
  when_expr: AnyParsedExpression | null;
  then_expr: AnyParsedExpression | null;
}
export interface CaseExpression extends ParsedExpression {
  case_checks: CaseCheck[];
  else_expr: AnyParsedExpression | null;
  class: "CASE";
}
export interface CastExpression extends ParsedExpression {
  child: AnyParsedExpression | null;
  cast_type: DuckDBLogicalType;
  try_cast: boolean;
  class: "CAST";
}
// ...
export interface CollateExpression extends ParsedExpression {
  child: AnyParsedExpression | null;
  collation: string;
  class: "COLLATE";
}
export interface ColumnBinding {
  table_index: number;
  column_index: number;
}
export interface ColumnDataRef extends TableRef {
  expected_names: string[];
  collection: unknown | null;
  type: "COLUMN_DATA";
}
// ...
export interface ColumnRefExpression extends ParsedExpression {
  column_names: string[];
  class: "COLUMN_REF";
}
export interface CommonTableExpressionInfo {
  aliases: string[];
  query: SelectStatement | null;
  materialized: string;
  key_targets: AnyParsedExpression[];
}
export interface CommonTableExpressionMap {
  map: Record<string, CommonTableExpressionInfo>;
}
export interface ComparisonExpression extends ParsedExpression {
  left: AnyParsedExpression | null;
  right: AnyParsedExpression | null;
  class: "COMPARISON";
}
export interface ConjunctionAndFilter extends TableFilter {
  child_filters: AnyTableFilter[];
  filter_type: "CONJUNCTION_AND";
}
export interface ConjunctionExpression extends ParsedExpression {
  children: AnyParsedExpression[];
  class: "CONJUNCTION";
}
export interface ConjunctionOrFilter extends TableFilter {
  child_filters: AnyTableFilter[];
  filter_type: "CONJUNCTION_OR";
}
export interface ConstantExpression extends ParsedExpression {
  value: DuckDBValue;
  class: "CONSTANT";
}
// ...
export interface DefaultExpression extends ParsedExpression {
  class: "DEFAULT";
}
// ...
export interface EmptyTableRef extends TableRef {
  type: "EMPTY_FROM";
}
// ...
export interface ExpressionListRef extends TableRef {
  expected_names: string[];
  expected_types: DuckDBLogicalType[];
  values: AnyParsedExpression[][];
  type: "EXPRESSION_LIST";
}
// ...
export interface FunctionExpression extends ParsedExpression {
  function_name: string;
  schema: string;
  children: AnyParsedExpression[];
  filter: AnyParsedExpression | null;
  order_bys: OrderModifier | null;
  distinct: boolean;
  is_operator: boolean;
  export_state: boolean;
  catalog: string;
  class: "FUNCTION";
}
// ...
export interface JoinRef extends TableRef {
  left: AnyTableRef | null;
  right: AnyTableRef | null;
  condition: AnyParsedExpression | null;
  join_type: string;
  ref_type: string;
  using_columns: string[];
  delim_flipped: boolean;
  duplicate_eliminated_columns: AnyParsedExpression[];
  /** @since v1.4.0 */
  is_implicit: boolean;
  type: "JOIN";
}
export interface LambdaExpression extends ParsedExpression {
  lhs: AnyParsedExpression | null;
  expr: AnyParsedExpression | null;
  /** @since v1.3.0 */
  syntax_type: string;
  class: "LAMBDA";
}
export interface LambdaRefExpression extends ParsedExpression {
  lambda_idx: number;
  column_name: string;
  class: "LAMBDA_REF";
}
// ...
export interface OperatorExpression extends ParsedExpression {
  children: AnyParsedExpression[];
  class: "OPERATOR";
}
export interface OptionalFilter extends TableFilter {
  child_filter: AnyTableFilter | null;
  filter_type: "OPTIONAL_FILTER";
}
export interface OrderByNode {
  type: string;
  null_order: string;
  expression: AnyParsedExpression | null;
}
export interface OrderModifier extends ResultModifier {
  orders: OrderByNode[];
  type: "ORDER_MODIFIER";
}
export interface ParameterExpression extends ParsedExpression {
  identifier: string;
  class: "PARAMETER";
}
export interface ParseInfo {
  info_type: string;
}
export interface ParsedExpression {
  class: string;
  type: string;
  alias: string;
  query_location: number;
}
export interface PivotColumn {
  pivot_expressions: AnyParsedExpression[];
  unpivot_names: string[];
  entries: PivotColumnEntry[];
  pivot_enum: string;
}
export interface PivotColumnEntry {
  values: DuckDBValue[];
  star_expr: AnyParsedExpression | null;
  alias: string;
}
export interface PivotRef extends TableRef {
  source: AnyTableRef | null;
  aggregates: AnyParsedExpression[];
  unpivot_names: string[];
  pivots: PivotColumn[];
  groups: string[];
  column_name_alias: string[];
  include_nulls: boolean;
  type: "PIVOT";
}
export interface PositionalReferenceExpression extends ParsedExpression {
  index: number;
  class: "POSITIONAL_REFERENCE";
}
export interface PragmaInfo extends ParseInfo {
  name: string;
  parameters: AnyParsedExpression[];
  named_parameters: Record<string, AnyParsedExpression>;
  info_type: "PRAGMA_INFO";
}
export interface QualifiedColumnName {
  catalog: string;
  schema: string;
  table: string;
  column: string;
}
export interface QueryNode {
  type: string;
  modifiers: AnyResultModifier[];
  cte_map: CommonTableExpressionMap;
}
export interface RecursiveCTENode extends QueryNode {
  cte_name: string;
  union_all: boolean;
  left: AnyQueryNode | null;
  right: AnyQueryNode | null;
  aliases: string[];
  key_targets: AnyParsedExpression[];
  type: "RECURSIVE_CTE_NODE";
}
// ...
export interface SampleOptions {
  sample_size: DuckDBValue;
  is_percentage: boolean;
  method: string;
  seed: number;
}
export interface ScalarMacroFunction extends MacroFunction {
  expression: AnyParsedExpression | null;
  type: "SCALAR_MACRO";
}
export interface SelectNode extends QueryNode {
  select_list: AnyParsedExpression[];
  from_table: AnyTableRef | null;
  where_clause: AnyParsedExpression | null;
  group_expressions: AnyParsedExpression[];
  group_sets: unknown[];
  aggregate_handling: string;
  having: AnyParsedExpression | null;
  sample: SampleOptions | null;
  qualify: AnyParsedExpression | null;
  type: "SELECT_NODE";
}
export interface SelectStatement {
  node: AnyQueryNode | null;
  named_param_map: Record<string, number>;
}
// ...
export interface SetOperationNode extends QueryNode {
  setop_type: string;
  left: AnyQueryNode | null;
  right: AnyQueryNode | null;
  setop_all: boolean;
  children: AnyQueryNode[];
  type: "SET_OPERATION_NODE";
}
// ...
export interface ShowRef extends TableRef {
  table_name: string;
  query: AnyQueryNode | null;
  show_type: string;
  catalog_name: string;
  schema_name: string;
  type: "SHOW_REF";
}
export interface StarExpression extends ParsedExpression {
  relation_name: string;
  exclude_list: string[];
  replace_list: Record<string, AnyParsedExpression>;
  columns: boolean;
  expr: AnyParsedExpression | null;
  unpacked: boolean;
  qualified_exclude_list: string[];
  rename_list: Record<string, string>;
  class: "STAR";
}
// ...
export interface SubqueryExpression extends ParsedExpression {
  subquery_type: string;
  subquery: SelectStatement | null;
  child: AnyParsedExpression | null;
  comparison_type: string;
  class: "SUBQUERY";
}
export interface SubqueryRef extends TableRef {
  subquery: SelectStatement | null;
  column_name_alias: string[];
  type: "SUBQUERY";
}
export interface TableColumn {
  name: string;
  type: DuckDBLogicalType;
}
export interface TableFilter {
  filter_type: string;
}
export interface TableFilterSet {
  filters: Record<string, AnyTableFilter>;
}
export interface TableFunctionRef extends TableRef {
  function: AnyParsedExpression | null;
  column_name_alias: string[];
  with_ordinality: string;
  type: "TABLE_FUNCTION";
}
export interface TableMacroFunction extends MacroFunction {
  query_node: AnyQueryNode | null;
  type: "TABLE_MACRO";
}
export interface TableRef {
  type: string;
  alias: string;
  sample: SampleOptions | null;
  query_location: number;
}
export interface TemplateTypeInfo extends ExtraTypeInfo {
  name: string;
  type: "TEMPLATE_TYPE_INFO";
}
export interface TransactionInfo extends ParseInfo {
  type: string;
  modifier: string;
  invalidation_policy: string;
  auto_rollback: boolean;
  info_type: "TRANSACTION_INFO";
}
export interface TypeExpression extends ParsedExpression {
  catalog: string;
  schema: string;
  type_name: string;
  children: AnyParsedExpression[];
  class: "TYPE";
}
// ...
export interface WindowExpression extends ParsedExpression {
  function_name: string;
  schema: string;
  catalog: string;
  children: AnyParsedExpression[];
  partitions: AnyParsedExpression[];
  orders: OrderByNode[];
  start: string;
  end: string;
  start_expr: AnyParsedExpression | null;
  end_expr: AnyParsedExpression | null;
  offset_expr: AnyParsedExpression | null;
  default_expr: AnyParsedExpression | null;
  ignore_nulls: boolean;
  filter_expr: AnyParsedExpression | null;
  exclude_clause: string;
  distinct: boolean;
  arg_orders: OrderByNode[];
  class: "WINDOW";
}
// ...
export type AnyParsedExpression = BetweenExpression | CaseExpression | CastExpression | CollateExpression | ColumnRefExpression | ComparisonExpression | ConjunctionExpression | ConstantExpression | DefaultExpression | FunctionExpression | LambdaExpression | LambdaRefExpression | OperatorExpression | ParameterExpression | PositionalReferenceExpression | StarExpression | SubqueryExpression | TypeExpression | WindowExpression;
export type AnyQueryNode = CTENode | RecursiveCTENode | SelectNode | SetOperationNode;
export type AnyResultModifier = DistinctModifier | LimitModifier | LimitPercentModifier | OrderModifier;
export type AnyTableFilter = ConjunctionAndFilter | ConjunctionOrFilter | ConstantFilter | DynamicFilter | ExpressionFilter | InFilter | IsNotNullFilter | IsNullFilter | OptionalFilter | StructFilter;
export type AnyTableRef = BaseTableRef | ColumnDataRef | EmptyTableRef | ExpressionListRef | JoinRef | PivotRef | ShowRef | SubqueryRef | TableFunctionRef;